metrics: Move the code a bit
This will allow adding subcommands for the metrics handler.
