+++ azafea/event_processors/metrics/v2/__init__.py
@@ -1,0 +1,2 @@
+# Reexport these so Azafea finds them
+from .handler import process  # noqa: F401
