+++ azafea/event_processors/metrics/v2/handler.py
@@ -12,8 +12,8 @@ import logging
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm.session import Session as DbSession
 
-from .events import new_aggregate_event, new_sequence_event, new_singular_event
-from .request import RequestBuilder
+from ..events import new_aggregate_event, new_sequence_event, new_singular_event
+from ..request import RequestBuilder
 
 
 log = logging.getLogger(__name__)
